+++ src/components/NavBar.module.css
@@ -6,22 +6,26 @@
   background-color: rgba(0, 0, 0, 0.4);
   backdrop-filter: blur(8px);
   -webkit-backdrop-filter: blur(8px);
-  padding: 4px 12px;
+  padding: 8px 12px;
   display: flex;
   justify-content: space-between;
   align-items: center;
   z-index: 1000;
-  height: 24px;
+  height: 32px;
   border-bottom: 1px solid rgba(255, 255, 255, 0.1);
   box-shadow: 0 2px 8px rgba(0, 0, 0, 0.15);
 }
 
-.navList {
-  list-style: none;
+.leftSection {
   display: flex;
-  gap: 10px;
-  padding: 0;
-  margin: 0;
+  align-items: center;
+  gap: 12px;
+}
+
+.iconContainer {
+  display: flex;
+  align-items: center;
+  gap: 12px;
 }
 
 .navLink {
@@ -38,10 +42,20 @@
   background-color: rgba(255, 255, 255, 0.1);
 }
 
-.iconContainer {
+/* Quick Start button styles */
+:global(.quickStartBtn) {
+  height: 24px;
+  padding: 0 12px;
+  border-radius: 4px;
+  font-size: 0.85rem;
+  font-weight: 500;
+  border: none;
+  cursor: pointer;
+  transition: all 0.2s ease;
+  white-space: nowrap;
   display: flex;
   align-items: center;
-  gap: 4px;
+  justify-content: center;
 }
 
 .iconWrapper {
@@ -70,4 +84,31 @@
 
 :global(.dark) .navLink:hover {
   background-color: rgba(255, 255, 255, 0.1);
+}
+
+.settingsButton {
+  color: #ffffff;
+  background: none;
+  border: none;
+  padding: 3px 6px;
+  border-radius: 2px;
+  transition: background-color 0.2s;
+  font-weight: 400;
+  font-size: 0.95rem;
+  cursor: pointer;
+  font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', 'Roboto', 'Oxygen',
+    'Ubuntu', 'Cantarell', 'Fira Sans', 'Droid Sans', 'Helvetica Neue',
+    sans-serif;
+}
+
+.settingsButton:hover {
+  background-color: rgba(255, 255, 255, 0.1);
+}
+
+:global(.light) .settingsButton {
+  color: #333333;
+}
+
+:global(.light) .settingsButton:hover {
+  background-color: rgba(0, 0, 0, 0.05);
 }
