+++ src/components/UnreleasedCard.module.css
@@ -3,31 +3,44 @@
   composes: card from './Card.module.css';
   opacity: 0.7;
   cursor: default;
+  padding: 16px;
 }
 
-.cardHeader {
-  composes: cardHeader from './Card.module.css';
-}
-
-.headerContent {
-  composes: headerContent from './Card.module.css';
-}
-
-.title {
-  composes: title from './Card.module.css';
-}
-
-.cardContent {
-  composes: cardContent from './Card.module.css';
+.content {
   display: flex;
   flex-direction: column;
   align-items: center;
   justify-content: center;
+  width: 100%;
+  gap: 12px;
+}
+
+.titleSection {
+  text-align: center;
+  width: 100%;
+}
+
+.title {
+  margin: 0;
+  font-size: var(--base-font-size);
+  line-height: 1.2;
+  font-weight: 600;
   text-align: center;
 }
 
+.descriptionSection {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  gap: 8px;
+  text-align: center;
+  width: 100%;
+}
+
 .description {
-  margin: 0 0 12px 0;
+  margin: 0;
+  font-size: var(--base-font-size);
 }
 
 .link {
@@ -37,6 +50,7 @@
   text-decoration: none;
   opacity: 0.7;
   transition: opacity 0.2s ease;
+  text-align: center;
 }
 
 .link:hover {
